Added Refuel and Occupancy Sensors
Refuel is having to appear as a Battery and not a propane tank because
the only "storage level" type of thing in HomeKit is the Battery. As a
result, the "Status Low Battery" flag will be set when the propane level
is below 15% or the Battery is below 25%. A future update will allow the
config file to set Low Percentage Levels if there is the need for that.
The Occupancy sensor tested is from the Ecobee3 interface.